+++ app/src/main/java/com/raulastete/todoistclone/domain/entity/Project.kt
@@ -1,9 +1,9 @@
 package com.raulastete.todoistclone.domain.entity
 
 data class Project(
-    val id: Int = 0,
+    val id: Long = 0L,
     val name: String,
-    val parentProjectId: Int = 0,
+    val parentProjectId: Long = 0L,
     val isFavorite: Boolean = false,
     val color: ProjectColor = ProjectColor.RED,
     val status: ProjectStatus = ProjectStatus.ACTIVE
